@@ -1,41 +1,45 @@
 [package]
 name = "insomnia-bot"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Evan Chang <evanc577@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [profile.release]
 lto = true
 
 [dependencies]
 anyhow = "1.0"
-approx = "0.5"
 if_chain = "1.0"
 futures = "0.3"
 markdown = "0.3"
-once_cell = "1.8"
-pyo3 = "0.15"
-regex = "1.5"
+once_cell = "1.14"
+poise = "0.3"
+pyo3 = "0.17"
+regex = "1.6"
 serde_json = "1.0"
 toml = "0.5"
-url = "2.2"
+url = "2.3"
 
 [dependencies.reqwest]
 version = "0.11"
 features = ["rustls-tls"]
 
 [dependencies.songbird]
-version = "~0.2.2"
+version = "0.3"
 features = ["builtin-queue", "yt-dlp"]
 
 [dependencies.serenity]
-version = "0.10"
-features = ["client", "standard_framework", "voice", "rustls_backend"]
+version = "0.11"
+default-features = false
+features = ["client", "voice", "rustls_backend"]
 
 [dependencies.serde]
 version = "1.0"
 features = ["derive"]
 
 [dependencies.tokio]
-version = "1.12"
+version = "1.21"
 features = ["macros", "rt-multi-thread", "signal", "time"]
+
+[dev-dependencies]
+approx = "0.5"
