@@ -5,30 +5,63 @@ mod music;
 
 use std::collections::HashMap;
 
-use crate::config::{Config, CONFIG_FILE};
-use crate::music::queue::QueueMutexMap;
-use crate::music::voice::CallMutexMap;
-use crate::music::{MUSIC_GROUP, MUSIC_HELP};
-
-use serenity::async_trait;
-use serenity::client::{Client, Context, EventHandler};
-use serenity::framework::StandardFramework;
-use serenity::model::gateway::Ready;
+use poise::serenity_prelude as serenity;
 use songbird::SerenityInit;
 use tokio::signal::unix::{signal, SignalKind};
 
-struct Handler;
+use crate::config::{Config, CONFIG_FILE};
+use crate::music::queue::QueueMutexMap;
+use crate::music::voice::CallMutexMap;
 
-#[async_trait]
-impl EventHandler for Handler {
-    async fn ready(&self, _ctx: Context, ready: Ready) {
-        println!("{} is connected!", ready.user.name);
+pub type Error = Box<dyn std::error::Error + Send + Sync>;
+pub type PoiseContext<'a> = poise::Context<'a, Data, Error>;
+#[derive(Debug)]
+pub struct Data {}
+
+/// Registers slash commands in this guild or globally
+#[poise::command(prefix_command, required_permissions = "MANAGE_GUILD", hide_in_help)]
+async fn register(ctx: PoiseContext<'_>) -> Result<(), Error> {
+    poise::builtins::register_application_commands_buttons(ctx).await?;
+    println!("Registering...");
+    Ok(())
+}
+
+/// Show this help menu
+#[poise::command(slash_command, prefix_command, track_edits)]
+async fn help(
+    ctx: PoiseContext<'_>,
+    #[description = "Specific command to show help about"]
+    #[autocomplete = "poise::builtins::autocomplete_command"]
+    command: Option<String>,
+) -> Result<(), Error> {
+    poise::builtins::help(
+        ctx,
+        command.as_deref(),
+        poise::builtins::HelpConfiguration {
+            show_context_menu_commands: true,
+            ..Default::default()
+        },
+    )
+    .await?;
+    Ok(())
+}
+
+async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
+    match error {
+        poise::FrameworkError::Setup { error, .. } => panic!("Failed to start bot: {:?}", error),
+        poise::FrameworkError::Command { error, ctx } => {
+            println!("Error in command `{}`: {:?}", ctx.command().name, error,);
+        }
+        error => {
+            if let Err(e) = poise::builtins::on_error(error).await {
+                println!("Error while handling error: {}", e)
+            }
+        }
     }
 }
 
 #[tokio::main]
 async fn main() {
-    // Configure the client with your Discord bot token in the environment.
     let config = match Config::get_config() {
         Ok(c) => c,
         Err(e) => {
@@ -37,22 +70,55 @@ async fn main() {
         }
     };
 
-    let framework = StandardFramework::new()
-        .configure(|c| {
-            c.prefix(&config.prefix);
-            c.case_insensitivity(true);
-            c
-        })
-        .help(&MUSIC_HELP)
-        .group(&MUSIC_GROUP);
+    // Add bot commands
+    let commands = vec![
+        register(),
+        help(),
+        music::album(),
+        music::list(),
+        music::pause(),
+        music::play(),
+        music::remove(),
+        music::skip(),
+        music::song(),
+        music::stop(),
+        music::video(),
+    ];
 
-    let mut client = Client::builder(&config.token)
-        .event_handler(Handler)
-        .framework(framework)
-        .register_songbird()
+    // Configure Poise options
+    let options = poise::FrameworkOptions {
+        commands,
+        prefix_options: poise::PrefixFrameworkOptions {
+            prefix: Some(config.prefix),
+            ..Default::default()
+        },
+        on_error: |error| Box::pin(on_error(error)),
+        pre_command: |ctx| {
+            Box::pin(async move {
+                println!(
+                    "command {} called by {}#{:04}",
+                    ctx.command().qualified_name,
+                    ctx.author().name,
+                    ctx.author().discriminator
+                );
+            })
+        },
+        ..Default::default()
+    };
+
+    let framework = poise::Framework::builder()
+        .options(options)
+        .client_settings(|cb| cb.register_songbird())
+        .token(config.token)
+        .intents(
+            serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT,
+        )
+        .user_data_setup(move |_ctx, _ready, _framework| Box::pin(async move { Ok(Data {}) }))
+        .build()
         .await
-        .expect("Err creating client");
+        .unwrap();
 
+    let mut client = framework.client();
     {
         let mut data = client.data.write().await;
         data.insert::<CallMutexMap>(HashMap::new());
